+++ server/ledger.mjs
@@ -13,3 +13,12 @@ export const getLedgerInfo = async (path, config) => {
         case "delegations": return await delegations(delegateKey)
     }
 }
+
+export const processGetDelegations = async () => {
+    const key = globalThis.config.publicKeyDelegators ?? globalThis.config.publicKey
+    let data = await delegations(key)
+
+    if (data) globalThis.nodeInfo.delegations = data
+
+    setTimeout(processGetDelegations, 600000)
+}